requirements.txt
pandasql
pandas
matplotlib
seaborn
numpy
streamlit
google-genai
python-dotenv
